+++ config/config.exs
@@ -1,10 +1,6 @@
 import Config
 
-config :alchemist_library, Library.Repo,
-  database: "postgres",
-  username: "postgres",
-  password: "1234",
-  hostname: "localhost"
-
 config :alchemist_library,
-      ecto_repos: [Library.Repo]
+  ecto_repos: [Library.Repo]
+
+import_config "#{config_env()}.exs"