+++ config/dev.exs
@@ -1,0 +1,7 @@
+import Config
+
+config :alchemist_library, Library.Repo,
+  database: "postgres",
+  username: "postgres",
+  password: "1234",
+  hostname: "localhost"