+++ lib/alchemist_library.ex
@@ -1,18 +1,2 @@
 defmodule AlchemistLibrary do
-  @moduledoc """
-  Documentation for `AlchemistLibrary`.
-  """
-
-  @doc """
-  Hello world.
-
-  ## Examples
-
-      iex> AlchemistLibrary.hello()
-      :world
-
-  """
-  def hello do
-    :world
-  end
 end