+++ lib/library/category.ex
@@ -1,9 +1,10 @@
 defmodule Library.Category do
   use Ecto.Schema
   import Ecto.Changeset
+  import Ecto.Query
 
   schema "categories" do
-    field :name, :string
+    field(:name, :string)
 
     timestamps()
   end
@@ -12,5 +13,48 @@ defmodule Library.Category do
     struct
     |> cast(params, [:name])
     |> validate_required([:name])
+    |> unique_constraint(:name)
+  end
+
+  def create_category(attrs) do
+    %__MODULE__{}
+    |> Library.Category.changeset(attrs)
+    |> Library.Repo.insert()
+  end
+
+  def get_all() do
+    query = from(Library.Category)
+    Library.Repo.all(query)
+  end
+
+  def get_by_name(name) do
+    Library.Repo.get_by(Library.Category, name: name)
+  end
+
+  def update_category(id, new_category) do
+    with %Library.Category{} = category <- Library.Repo.get(Library.Category, id),
+         changeset = Library.Category.changeset(category, new_category),
+         {:ok, updated_category} <- Library.Repo.update(changeset) do
+      {:ok, updated_category}
+    else
+      nil ->
+        {:error, :not_found}
+
+      {:error, changeset} ->
+        {:error, changeset}
+    end
+  end
+
+  def delete_category(id) do
+    with %Library.Category{} = category <- Library.Repo.get(Library.Category, id),
+         {:ok, deleted_category} <- Library.Repo.delete(category) do
+      {:ok, deleted_category}
+    else
+      nil ->
+        {:error, :not_found}
+
+      {:error, changeset} ->
+        {:error, changeset}
+    end
   end
 end