+++ priv/repo/migrations/20250505124306_category_name_unique.exs
@@ -1,0 +1,7 @@
+defmodule Library.Repo.Migrations.CategoryNameUnique do
+  use Ecto.Migration
+
+  def change do
+    create unique_index(:categories, [:name])
+  end
+end