+++ test/alchemist_library_test.exs
@@ -1,8 +1,2 @@
 defmodule AlchemistLibraryTest do
-  use ExUnit.Case
-  doctest AlchemistLibrary
-
-  test "greets the world" do
-    assert AlchemistLibrary.hello() == :world
-  end
 end