+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add Books, Authors and Categories modules
 - Create Books Migration
 - Create Author Migration
 - Create Category Migration